Speed-up MS dataset reading
... by specifying dtype beforehand. All values are numeric (though some of
them are ints, but that does not matter for further analysis).

+++ ms_datasets.py
@@ -54,7 +54,7 @@ def add_deltas_and_sanitize_time(dataset: pd.DataFrame, delta_quantities: Sequen
 def prepare_sampled_merged_data(
         path: str = 'C:/MyData/Versetzungsdaten/Voxel_Data/sampled_merged_last_voxel_data_size2400_order2_speedUp2.csv',
         delta_steps: int = 1, subset: str = 'none') -> pd.DataFrame:
-    dataset = pd.read_csv(path)
+    dataset = pd.read_csv(path, dtype='float64')  # specifying dtype makes reading faster
     dataset.drop(columns=dataset.columns[0], inplace=True)  # drop 1st column (unnamed id column)
     # String matching functions are more difficult to use if name of one reaction type is substring
     # of another one, thus we rename "multiple_coll" to avoid clashes with "coll"
@@ -70,7 +70,7 @@ def prepare_sampled_merged_data(
 def prepare_sampled_voxel_data(
         path: str = 'C:/MyData/Versetzungsdaten/Voxel_Data/sampled_voxel_data_size2400_order2_speedUp2.csv',
         delta_steps: int = 1, subset: str = 'none') -> pd.DataFrame:
-    dataset = pd.read_csv(path)
+    dataset = pd.read_csv(path, dtype='float64')  # specifying dtype makes reading faster
     dataset.drop(columns=dataset.columns[0], inplace=True)  # drop 1st column (unnamed id column)
     # In current datasets, slip systems notation for shear is "shear(1)" instead "shear_gs(1)"
     dataset.rename(columns=lambda x: re.sub('shear', 'shear_gs', x), inplace=True)
